+ chunking logic

+++ embedding/embedd.py
@@ -16,7 +16,10 @@ MODEL='embeddinggemma:300m'
 import ollama
 import chromadb
 import os 
-os.system(r'rm -rf embedded_data/chroma_db')
+
+from glob import glob
+if glob(r'embedded_data/chroma_db'):
+    os.system(r'rm -rf embedded_data/chroma_db')
 
 # Start client (creates local .chroma DB folder)
 client = chromadb.PersistentClient(path=r'./embedded_data/chroma_db')
@@ -24,7 +27,7 @@ client = chromadb.PersistentClient(path=r'./embedded_data/chroma_db')
 # should be a variable
 collection = client.create_collection('DMAS')
 import json 
-with open(r'raw_data/meta_data.json', 'r') as meta_data_fp:
+with open(r'raw_data/DMAS_CHUNK_META.json', 'r') as meta_data_fp:
     meta_data = json.load(meta_data_fp)
 
 # Insert into DB
@@ -32,12 +35,15 @@ from tqdm import tqdm
 page_idx = 0
 for page in tqdm(meta_data):
     chunk_idx = 0
-    for chunk in page['chunks']:
-        filename = chunk['chunk_data_filename']
-        chunk_path = r'raw_data/chunks/' + filename
-        with open(chunk_path) as cf:
-            text = cf.read()
-            text_embedding = ollama.embeddings(model=MODEL, prompt=text)["embedding"]
-            collection.add(documents=[text], embeddings=[text_embedding], ids=[f"{page_idx}/{chunk_idx}"])
-        chunk_idx += 1
+    # not every page needs to be embedded
+    if 'chunks' in page:
+        for chunk in page['chunks']:
+            filename = chunk['chunk_data_filename']
+            chunk_path = r'raw_data/chunks/' + filename
+            with open(chunk_path) as cf:
+                text = cf.read()
+                text_embedding = ollama.embeddings(model=MODEL, prompt=text)["embedding"]
+                # collecttion meta data does not take lists, so we are not using it
+                collection.add(documents=[text], embeddings=[text_embedding], ids=[f"{page_idx}/{chunk_idx}"])
+            chunk_idx += 1
     page_idx += 1
